chore: create external new project page
